added usePannign hook, fixed useRenderer lint issue
